moved error to seperate file, regrouped imports, removed extra copyright notice
